+++ web-app/packages/app/src/modules/project/views/ProjectSettingsView.vue
@@ -9,20 +9,14 @@ SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
     :namespace="namespace"
     :projectName="projectName"
     :show-settings="projectStore.isProjectOwner"
-    :show-access-requests="userStore.isGlobalWorkspaceAdmin"
   >
   </project-settings-view-template>
 </template>
 
 <script setup lang="ts">
-import {
-  ProjectSettingsViewTemplate,
-  useProjectStore,
-  useUserStore
-} from '@mergin/lib'
+import { ProjectSettingsViewTemplate, useProjectStore } from '@mergin/lib'
 
 defineProps<{ namespace: string; projectName: string }>()
 
 const projectStore = useProjectStore()
-const userStore = useUserStore()
 </script>
