+++ web-app/packages/lib/src/modules/project/views/ProjectCollaboratorsViewTemplate.vue
@@ -6,6 +6,18 @@ SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
 
 <template>
   <div>
+    <app-container
+      v-if="showAccessRequests && projectStore.accessRequestsCount > 0"
+    >
+      <app-section
+        ><template #title
+          >Requests
+          <span class="opacity-80"
+            >({{ projectStore.accessRequestsCount }})</span
+          ></template
+        ><project-access-requests
+      /></app-section>
+    </app-container>
     <!-- Searching -->
     <app-container>
       <app-section ground>
@@ -39,18 +51,25 @@ SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
 
 <script setup lang="ts">
 import { MenuItem, MenuItemCommandEvent } from 'primevue/menuitem'
-import { computed } from 'vue'
+import { computed, watchEffect } from 'vue'
 
 import { ProjectMembersTable } from '../components'
+import ProjectAccessRequests from '../components/ProjectAccessRequests.vue'
 import { useProjectStore } from '../store'
 
 import AppContainer from '@/common/components/AppContainer.vue'
 import AppMenu from '@/common/components/AppMenu.vue'
 import AppSection from '@/common/components/AppSection.vue'
+import { useUserStore } from '@/main'
 
 defineEmits<{ share: [] }>()
+withDefaults(defineProps<{ showAccessRequests?: boolean }>(), {
+  showAccessRequests: false
+})
 
 const projectStore = useProjectStore()
+const userStore = useUserStore()
+
 const filterMenuItems = computed<MenuItem[]>(() =>
   [
     {
@@ -80,6 +99,16 @@ const filterMenuItems = computed<MenuItem[]>(() =>
         : ''
   }))
 )
+
+// Every dashboard table have to load data on orkspace change
+watchEffect(() => {
+  if (userStore.isWorkspaceAdmin()) {
+    projectStore.initNamespaceAccessRequests({
+      namespace: userStore.currentWorkspace.name,
+      params: null
+    })
+  }
+})
 </script>
 
 <style lang="scss" scoped></style>
